refactory and add setup.py
